Redo: Changes board.c's variables to short

+++ src/board.h
@@ -5,11 +5,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-Mask shift(const size_t row, const size_t col);
-size_t pos2Shift(const size_t row, const size_t col);
+Mask shift(const short row, const short col);
+short pos2Shift(const short row, const short col);
 void copyNode(Node *destination, const Node *origin);
 void copyBoard(Board *destination, const Board *origin);
-int makeMove(Node *node, const size_t col);
+short makeMove(Node *node, const short col);
 void printBoard(const Node *node);
 void printNode(const Node *node);
